requirement.txt
ffmpeg-python
audio_separator
pyannote.audio
python-dotenv
pydub 
python-whisper
transformers
librosa